feat: 페이지 접근 제한을 위해 hoc 도입
